requirements.txt
# LLM Providers (Gemini primary, Claude optional)
google-generativeai>=0.8.0
anthropic>=0.40.0
elevenlabs>=1.0.0
ffmpeg-python>=0.2.0
fastmcp>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
matplotlib>=3.8.0
requests>=2.31.0
PyYAML>=6.0.1
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
